pyproject.toml
[project]
name = "detector-backend"
version = "1.0.0"
description = "Logo Detection API with GPU acceleration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "python-multipart>=0.0.20",
    "opencv-python>=4.12.0.88",
    "ultralytics>=8.3.170",
    "pillow>=11.3.0",
    "numpy>=2.2.6",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.4",
]

[project.optional-dependencies]
dev = [
    "requests>=2.31.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
start = "main:main"
dev = "main:dev"
test = "pytest:main"
